Adds comment functionality
- add comment model, migration, and relationships
- add comment service for business logic and authorization
- add comment controller with api endpoints for managing comments
- add form requests for comment validation
- add api resource for comment representation
- add api routes for comment management
- adjust integer types in existing migrations

+++ database/migrations/2025_05_29_180916_create_books_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("title", 50);
             $table->string("image_url", 100)->nullable();
             $table->string("author", 50);
-            $table->unsignedSmallInteger("year")->nullable();
+            $table->integer("year")->nullable();
             $table->string("description", 500)->nullable();
             $table->string("tags", 100)->nullable();
             $table->boolean("public")->default(true);
@@ -50,7 +50,7 @@ return new class extends Migration
         Schema::create("bookmarks", function (Blueprint $table) {
             $table->id();
             $table->string("name", 20)->nullable();
-            $table->unsignedSmallInteger("page");
+            $table->smallInteger("page");
             $table->timestamp("created_at");
             $table->foreignId("book_id")->constrained()->cascadeOnDelete()->cascadeOnUpdate();
         });
